Make gc test more reliable
