+++ jest.gc.config.json
@@ -1,0 +1,6 @@
+{
+  "transform": {
+    "^.+\\.(t|j)s?$": "ts-jest"
+  },
+  "moduleFileExtensions": ["ts", "js", "json"]
+}
