+++ src/thread-pool.spec.ts
@@ -407,38 +407,6 @@ describe('Callbacks', () => {
     expect(callback).toHaveBeenCalledWith(3)
   })
 
-  // Note: This might be flaky, as the garbage collection and the finalizer cannot be triggered reliably
-  it('cleans up main thread function when garbage collected on thread', async () => {
-    const worker = await createThreadPool<WorkerWithCallback>('./__tests__/workers/callback')
-    // Need to create a lot of callbacks to trigger the garbage collection
-    const callTimes = 50000
-    const msgHandler = jest.fn()
-
-    worker.pool.on('thread:message', (msg) => {
-      if(isThreadFreeFunctionMessage(msg)) {
-        msgHandler(msg)
-      }
-    })
-
-    const callback = jest.fn()
-    for (let i = 0; i < callTimes; i++) {
-      await worker.withCallback(1, 2, callback)
-    }
-    
-    await new Promise<void>((resolve) => setTimeout(() => resolve(), 10000))
-    worker.pool.terminate()
-
-    expect(callback.mock.calls.length).toBeGreaterThan(10000)
-    expect(callback).toHaveBeenCalledWith(3)
-    expect(msgHandler).toHaveBeenCalledWith({
-      action: ThreadMessageAction.FREE_FUNCTION,
-      functionId: expect.any(Number),
-      key: expect.any(String)
-    })
-    const numberOfStoredMethods = worker.pool.callbacks.get('withCallback')?.size
-    expect(numberOfStoredMethods).toBeLessThan(callTimes / 2)
-  }, 20000)
-
   it('can call a callback function from all threads', async () => {
     const worker = await createThreadPool<WorkerWithCallback>('./__tests__/workers/callback', {
       size: 2
